requirements.txt
python-can
pyserial
PyYAML
pybullet
flask-cors
requests